# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de gestion des comptes bancaires",
        "description": "API pour gérer les comptes bancaires",
        "contact": {
            "name": "Faye Yatedene",
            "email": "faye.yatedene@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://ges-compte.onrender.com/api/v1",
            "description": "Serveur de production"
        },
        {
            "url": "http://127.0.0.1:3000/api",
            "description": "Serveur de développement"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/v1/authentication/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion utilisateur",
                "description": "Authentifie un utilisateur et renvoie un jeton d'accès",
                "operationId": "f8b21f278a93ec5f570450985b0f54e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "UNAUTHORIZED"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Identifiants invalides"
                                                },
                                                "timestamp": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "traceId": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/authentication/refresh": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Actualiser le jeton d'accès",
                "description": "Génère un nouveau jeton d'accès en utilisant le jeton d'actualisation",
                "operationId": "e2dbdd2167289a88d167618805ced1ad",
                "responses": {
                    "200": {
                        "description": "Jeton actualisé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Jeton d'actualisation manquant ou invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "UNAUTHORIZED"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Jeton d'actualisation manquant"
                                                },
                                                "timestamp": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "traceId": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/authentication/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion utilisateur",
                "description": "Révoque le jeton d'accès de l'utilisateur actuel",
                "operationId": "39619ff291ed512ab58391ed7bbfdb01",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Déconnecté"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "UNAUTHORIZED"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Unauthorized"
                                                },
                                                "timestamp": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "traceId": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "AuthResponse": {
                "properties": {
                    "access_token": {
                        "type": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9..."
                    },
                    "token_type": {
                        "type": "string",
                        "example": "Bearer"
                    },
                    "expires_in": {
                        "type": "integer",
                        "example": 3600
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}